# docs/swagger.yaml
definitions:
  request.ChangePasswordReq:
    properties:
      newPassword:
        description: 新密码
        type: string
      password:
        description: 密码
        type: string
    type: object
  request.Login:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
  request.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    required:
    - page
    - pageSize
    type: object
  request.Register:
    properties:
      email:
        type: string
      nickName:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.SessionDetailReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  request.SessionListReq:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      type:
        description: 1:hight bonus 2:hight rate 3:hot"
        type: integer
    required:
    - page
    - pageSize
    - type
    type: object
  response.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/response.SysUserResponse'
    type: object
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  response.SysUserResponse:
    properties:
      balance:
        type: number
      email:
        type: string
      freezeBalance:
        type: number
      nickName:
        type: string
      phone:
        type: string
      uid:
        type: integer
      userName:
        type: string
    type: object
  system.ActivitySession:
    properties:
      activityBonus:
        type: integer
      activityId:
        type: integer
      activityLimitCount:
        type: integer
      activityName:
        type: string
      activitySpend:
        type: integer
      id:
        type: integer
      openTime:
        type: integer
      status:
        type: integer
      uids:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/change/password:
    post:
      parameters:
      - description: 原密码, 新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 用户修改密码
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户中心
  /api/deposit/history:
    get:
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 充值记录
      tags:
      - 用户中心
  /api/game/history:
    post:
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: 抽奖记录
      tags:
      - 活动中心
  /api/home/recommand:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  $ref: '#/definitions/system.ActivitySession'
              type: object
      summary: 首页推荐
      tags:
      - 活动中心
  /api/login:
    post:
      parameters:
      - description: 用户名, 密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 返回包括用户信息,token
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResponse'
                msg:
                  type: string
              type: object
      summary: 登录
      tags:
      - 用户中心
  /api/register:
    post:
      parameters:
      - description: 用户名, 昵称, 密码,
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: 用户注册账号,返回包括用户信息 token
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResponse'
                msg:
                  type: string
              type: object
      summary: 注册账号
      tags:
      - 用户中心
  /api/session/check:
    post:
      parameters:
      - description: 场次ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SessionDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: 是否已购入场券
      tags:
      - 活动中心
  /api/session/detail:
    post:
      parameters:
      - description: 场次ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SessionDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  $ref: '#/definitions/system.ActivitySession'
              type: object
      summary: 活动场次详情
      tags:
      - 活动中心
  /api/session/list:
    post:
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SessionListReq'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: 场次列表
      tags:
      - 活动中心
  /api/session/ticket:
    post:
      parameters:
      - description: 场次ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SessionDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  $ref: '#/definitions/system.ActivitySession'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 购买入场券
      tags:
      - 活动中心
  /api/user/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 返回包括用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  $ref: '#/definitions/response.LoginResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户信息
      tags:
      - 用户中心
  /api/withdraw/history:
    get:
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 提现记录
      tags:
      - 用户中心
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"